tests/wpt/tests/html/semantics/permission-element/bounded-css-properties.tentative.html

<!DOCTYPE html>
<meta charset=utf-8>
<link rel="help" href="https://github.com/WICG/PEPC/blob/main/explainer.md#locking-the-pepc-style">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<!--The permission element should have some limits for specific properties:
  * font-weight is adjusted to be at least 200.
  * font-style should only have "normal" or "italic" values.
  * word-spacing should be at most 0.5 of the font size, and non-negative.
-->
<style>
  #id1 {
    font-weight: 100;
    font-style: oblique 30deg;
    word-spacing: 1em;
    font-size: 100px;
  }
  #id2 {
    font-weight: 300;
    font-style: italic;
    word-spacing: 0.4em;
    font-size: 100px;
  }
  #id3 {
    word-spacing: -1px;
  }
</style>


<permission id="id1" type="geolocation">
<permission id="id2" type="camera">
<permission id="id3" type="microphone">

<script>
  test(function(){
    var el_outside_bounds = document.getElementById("id1");
    assert_equals(getComputedStyle(el_outside_bounds).fontWeight, "200", "font-weight");
    assert_equals(getComputedStyle(el_outside_bounds).fontStyle, "normal", "font-style");
    assert_equals(getComputedStyle(el_outside_bounds).wordSpacing, "50px", "word-spacing");

    var el_negative_bounds = document.getElementById("id3");
    assert_equals(getComputedStyle(el_negative_bounds).wordSpacing, "0px", "word-spacing, negative");
  }, "Properties with out-of-bounds values should be corrected");

  test(function(){
    var el_inside_bounds = document.getElementById("id2");
    assert_equals(getComputedStyle(el_inside_bounds).fontWeight, "300", "font-weight");
    assert_equals(getComputedStyle(el_inside_bounds).fontStyle, "italic", "font-style");
    assert_equals(getComputedStyle(el_inside_bounds).wordSpacing, "40px", "word-spacing");
  }, "Properties with values in bounds should not be modified");
</script>
</body>
